# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Librarr API",
    "description": "Self-hosted book, audiobook, and manga search and download manager \u2014 the *arr for books. The HTTP/JSON API powers both the SPA and any external tooling (CLI scripts, AI agents, *arr-family integrations).",
    "version": "1.x",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5050",
      "description": "Default local install"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "API key for programmatic access. Also accepted via ?apikey= query parameter."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "librarr_session",
        "description": "Set by POST /api/login. Used by the SPA."
      },
      "opdsBasic": {
        "type": "http",
        "scheme": "basic",
        "description": "HTTP Basic authentication for OPDS reader apps using existing enabled Librarr users."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "WantedBook": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "isbn": {
            "type": "string"
          },
          "asin": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "media_type": {
            "type": "string"
          },
          "preferred_format": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "origin_source": {
            "type": "string"
          },
          "origin_release_title": {
            "type": "string"
          },
          "origin_indexer": {
            "type": "string"
          },
          "source_id": {
            "type": "string"
          },
          "monitored": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          },
          "selected_release_id": {
            "type": "integer"
          },
          "selected_release_title": {
            "type": "string"
          },
          "download_job_id": {
            "type": "string"
          },
          "download_client": {
            "type": "string"
          },
          "download_hash": {
            "type": "string"
          },
          "download_started_at": {
            "type": "string",
            "format": "date-time"
          },
          "download_error": {
            "type": "string"
          }
        }
      },
...
      "WantedReleaseDownloadResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "wanted_id": {
            "type": "integer"
          },
          "release_id": {
            "type": "integer"
          },
          "download_job_id": {
            "type": "string"
          },
          "download_hash": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "warning": {
            "type": "string"
          },
          "item": {
            "$ref": "#/components/schemas/WantedBook"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "apiKey": []
    },
    {
      "session": []
    }
  ],
  "paths": {
...
    "/torznab/api": {
      "get": {
        "summary": "Torznab indexer API",
        "description": "Lets Prowlarr / Readarr / other *arr apps use Librarr as a Torznab indexer.",
        "parameters": [
          {
            "name": "t",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "caps",
                "search",
                "book",
                "audio"
              ]
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "title",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "author",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apikey",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Required when TORZNAB_API_KEY is set."
          }
        ],
        "responses": {
          "200": {
            "description": "Torznab XML (RSS feed or caps document)",
            "content": {
              "application/xml": {}
            }
          }
        }
      }
    },
...
    "/opds/download/{id}": {
      "get": {
        "summary": "Download a library file for OPDS acquisition",
        "description": "Streams a normalized library file by stable file ID without exposing filesystem paths.",
        "security": [
          {
            "opdsBasic": []
          },
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Library file stream"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "File not found"
          }
        }
      }
    },
...
    "/api/v1/wanted/{id}/releases/{release_id}/download": {
      "post": {
        "summary": "Send a stored wanted release to the configured torrent client",
        "description": "Manually submits one stored Wanted release to the existing torrent download pipeline. The client sends only the Wanted id and stored release id; internal download URLs remain server-side.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Release submitted to the torrent client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WantedReleaseDownloadResponse"
                }
              }
            }
          },
          "404": {
            "description": "Wanted book or release not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Wanted item is already downloading or complete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Release cannot be submitted through the configured torrent pipeline",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Torrent client handoff failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}